server/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-multipart
black
python-dotenv
rasterio
earthengine-api
geemap
geojson
